Allow script to make the commit
Previously, the script would need to stage changes but could not commit them. This PR allows the script to make the commit if the title and body are omitted from the command. In this case, the script is going to need to 'git add' and 'git commit' the changes. The advantage of this is that the script can write the commit message, which allows you to have a dynamically generated commit message.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>